LPS-95357 Extract methods to Util
